requirements.txt
fastapi[standard]
passlib[bcrypt]
bcrypt==4.0.1
pytest
sqlmodel
psycopg2-binary
python-jose
python-dotenv